print points without board, line, lines class
