+++ src/main/java/Controller/Calculator.java
@@ -1,13 +1,11 @@
 package Controller;
 
-import Model.Board;
 import Model.Figure;
 import View.InputView;
 import java.util.ArrayList;
 
 public class Calculator {
 
-    private static Board board = Board.of();
     private static Figure figure = new Figure();
 
     public void initPoints() {
@@ -16,7 +14,6 @@ public class Calculator {
             System.out.println("Input coordinates. ex. (1,2)-(3,4)-(5,1)");
             coordinates = Utils.checkReturnCoordinates(InputView.initCoordinate());
             figure.setCoordinates(coordinates);
-            board.setFigure(figure);
         } catch (RuntimeException e) {
             initPoints();
         }
@@ -26,7 +23,8 @@ public class Calculator {
         figure.calculateArea();
     }
 
-    public Board getBoard() {
-        return board;
+    public Figure getFigure() {
+        return figure;
     }
+
 }
