+++ src/main/java/View/ResultView.java
@@ -11,10 +11,10 @@ import java.util.stream.IntStream;
 public class ResultView {
 
     public static void printBoard(Calculator controller) {
-        Board board = controller.getBoard();
+        Figure figure = controller.getFigure();
         for (int y = 23; y >= 0; y--) {
             print(drawYaxis(y));
-            print(drawLine(y, board));
+            print(drawLine(y, figure));
             System.out.println();
         }
         drawXaxis();
@@ -29,14 +29,14 @@ public class ResultView {
         return "      |";
     }
 
-    static String drawPoint(int x, int y, Board board) {
-        if (board.isEmpty(x, y)) return "  ";
-        return board.getPoint(x, y).toString();
+    static String drawPoint(int x, int y, Figure figure) {
+        if (figure.isAtCoordinate(x, y)) return Point.of(x, y).toString();
+        return "  ";
     }
 
     // yet not perfect logic
-    static String drawLine(int y, Board board) {
-        return IntStream.range(0, 24).mapToObj(i -> drawPoint(i, y, board)).collect(Collectors.joining(""));
+    static String drawLine(int y, Figure figure) {
+        return IntStream.range(0, 24).mapToObj(i -> drawPoint(i, y, figure)).collect(Collectors.joining(""));
     }
 
     static String drawXaxisLine() {
